[refactor] 패키지 구조 및 이름 변경
