+++ feature/home/src/main/java/com/tasker/android/home/presentation/home/calendar_view/HomeCalendarAdapter.kt
@@ -1,4 +1,4 @@
-package com.tasker.android.home.presentation.calendar
+package com.tasker.android.home.presentation.home.calendar_view
 
 import android.annotation.SuppressLint
 import android.content.Context
