+++ feature/home/src/main/java/com/tasker/android/home/presentation/home/calendar_view/HomeCalendarBottomSheetViewModel.kt
@@ -1,4 +1,4 @@
-package com.tasker.android.home.presentation.calendar
+package com.tasker.android.home.presentation.home.calendar_view
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
