+++ feature/home/src/main/java/com/tasker/android/home/presentation/home/category/HomeCategoryViewFragment.kt
@@ -1,4 +1,4 @@
-package com.tasker.android.home.presentation.time
+package com.tasker.android.home.presentation.home.category
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,13 +7,13 @@ import android.view.View
 import android.view.ViewGroup
 import com.tasker.android.home.R
 
-class HomeTimeViewFragment : Fragment() {
+class HomeCategoryViewFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home_time_type, container, false)
+        return inflater.inflate(R.layout.fragment_home_category_type, container, false)
     }
 }
