+++ feature/home/src/main/java/com/tasker/android/home/presentation/home/date_picker/HomeDatePickerItemDecoration.kt
@@ -1,4 +1,4 @@
-package com.tasker.android.home.presentation.date_picker
+package com.tasker.android.home.presentation.home.date_picker
 
 import android.content.Context
 import android.graphics.Rect
