+++ feature/home/src/main/java/com/tasker/android/home/presentation/home/main/HomeTaskViewAdapter.kt
@@ -1,10 +1,10 @@
-package com.tasker.android.home.presentation.main
+package com.tasker.android.home.presentation.home.main
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.tasker.android.home.presentation.time.HomeTimeViewFragment
-import com.tasker.android.home.presentation.category.HomeCategoryViewFragment
-import com.tasker.android.home.presentation.list.HomeListViewFragment
+import com.tasker.android.home.presentation.home.time_view.HomeTimeViewFragment
+import com.tasker.android.home.presentation.home.category.HomeCategoryViewFragment
+import com.tasker.android.home.presentation.home.list_view.HomeListViewFragment
 
 class HomeTaskViewAdapter(fragment: Fragment) : FragmentStateAdapter(fragment) {
     override fun getItemCount(): Int = 3
