+++ feature/home/src/main/java/com/tasker/android/home/presentation/home/time_view/HomeTimeViewFragment.kt
@@ -1,4 +1,4 @@
-package com.tasker.android.home.presentation.category
+package com.tasker.android.home.presentation.home.time_view
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,13 +7,13 @@ import android.view.View
 import android.view.ViewGroup
 import com.tasker.android.home.R
 
-class HomeCategoryViewFragment : Fragment() {
+class HomeTimeViewFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home_category_type, container, false)
+        return inflater.inflate(R.layout.fragment_home_time_type, container, false)
     }
 }
