+++ feature/home/src/main/res/layout/fragment_home_calendar_bottom_sheet.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".presentation.calendar.HomeCalendarBottomSheetFragment">
+    tools:context=".presentation.home.calendar_view.HomeCalendarBottomSheetFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
