+++ feature/home/src/main/res/layout/fragment_home_category_type.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.category.HomeCategoryViewFragment">
+    tools:context=".presentation.home.category.HomeCategoryViewFragment">
 
     <TextView
         android:layout_width="match_parent"
