+++ feature/home/src/main/res/layout/fragment_home_list_view.xml
@@ -6,13 +6,13 @@
     <data>
         <variable
             name="viewModel"
-            type="com.tasker.android.home.presentation.list.HomeListViewModel" />
+            type="com.tasker.android.home.presentation.home.list_view.HomeListViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.list.HomeListViewFragment">
+        tools:context=".presentation.home.list_view.HomeListViewFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv_home_list_view"
