+++ feature/home/src/main/res/layout/fragment_home_time_type.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.time.HomeTimeViewFragment">
+    tools:context=".presentation.home.time_view.HomeTimeViewFragment">
 
     <TextView
         android:layout_width="match_parent"
